requirements.txt
mysql-connector-python==8.0.28
Unidecode==1.3.4
numpy==1.21.6